pyproject.toml
[tool.poetry]
name = "azure-app-registration-monitor"
version = "0.1.0"
description = ""
authors = ["githubofkrishnadhas <krishnadhasnk1997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
pytz = "^2024.1"
msal = "^1.29.0"
requests = "^2.32.3"
python-dateutil = "^2.9.0.post0"
sendgrid = "^6.11.0"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
